+++ UnityEngine.TestRunner/NUnitExtensions/Commands/BeforeAfterTestCommandBase.cs
@@ -97,12 +97,9 @@ namespace UnityEngine.TestTools
         {
             var unityContext = (UnityTestExecutionContext)context;
             var state = GetState(unityContext);
-
-            // When entering PlayMode state will incorrectly be seen as null. Looking at the hashcode to be certain that it is null.
-            if (state?.GetHashCode() == null)
+            if (state == null)
             {
-                // We do not expect a state to exist in playmode
-                state = ScriptableObject.CreateInstance<BeforeAfterTestCommandState>();
+                throw new Exception($"No state in context for {GetType().Name}.");
             }
 
             if(state.ShouldRestore)