+++ UnityEngine.TestRunner/NUnitExtensions/Commands/TestActionCommand.cs
@@ -42,7 +42,8 @@ namespace UnityEngine.TestTools
 
         protected override BeforeAfterTestCommandState GetState(UnityTestExecutionContext context)
         {
-            return null;
+            // TestActionCommand does not support domain reloads and will not need a persisted state.
+            return new BeforeAfterTestCommandState();
         }
     }
 }