+++ UnityEngine.TestRunner/NUnitExtensions/TestExtensions.cs
@@ -10,17 +10,14 @@ namespace UnityEngine.TestRunner.NUnitExtensions
 {
     internal static class TestExtensions
     {
-        private static IEnumerable<string> GetTestCategories(this ITest test)
+        private static IEnumerable<string> GetCategoriesInTestAndAncestors(ITest test)
         {
             var categories = test.Properties[PropertyNames.Category].Cast<string>().ToList();
-            if (categories.Count == 0 && test is TestMethod)
+            if (test.Parent != null)
             {
-                // only mark tests as Uncategorized if the test fixture doesn't have a category,
-                // otherwise the test inherits the Fixture category
-                var fixtureCategories = test.Parent.Properties[PropertyNames.Category].Cast<string>().ToList();
-                if (fixtureCategories.Count == 0)
-                    categories.Add(CategoryFilterExtended.k_DefaultCategory);
+                categories.AddRange(GetCategoriesInTestAndAncestors(test.Parent));
             }
+
             return categories;
         }
 
@@ -32,11 +29,13 @@ namespace UnityEngine.TestRunner.NUnitExtensions
 
         public static List<string> GetAllCategoriesFromTest(this ITest test)
         {
-            if (test.Parent == null)
-                return test.GetTestCategories().ToList();
-
-            var categories = GetAllCategoriesFromTest(test.Parent);
-            categories.AddRange(test.GetTestCategories());
+            var categories = GetCategoriesInTestAndAncestors(test).ToList();
+            if (categories.Count == 0 && test is TestMethod)
+            {
+                // only mark tests as Uncategorized if the test fixture doesn't have a category,
+                // otherwise the test inherits the Fixture category
+                categories.Add(CategoryFilterExtended.k_DefaultCategory);
+            }
             return categories;
         }