+++ UnityEngine.TestRunner/TestRunner/PlaymodeTestsController.cs
@@ -93,8 +93,11 @@ namespace UnityEngine.TestTools.TestRunner
                 yield return null;
             }
 
+            var context = new UnityTestExecutionContext();
+            UnityTestExecutionContext.CurrentContext = context;
+
             var testListUtil = new PlayerTestAssemblyProvider(new AssemblyLoadProxy(), m_AssembliesWithTests);
-            m_Runner = new UnityTestAssemblyRunner(new UnityTestAssemblyBuilder(settings.orderedTestNames), new PlaymodeWorkItemFactory());
+            m_Runner = new UnityTestAssemblyRunner(new UnityTestAssemblyBuilder(settings.orderedTestNames), new PlaymodeWorkItemFactory(), context);
 
             var loadedTests = m_Runner.Load(testListUtil.GetUserAssemblies().Select(a => a.Assembly).ToArray(), TestPlatform.PlayMode, UnityTestAssemblyBuilder.GetNUnitTestBuilderSettings(TestPlatform.PlayMode));
             loadedTests.ParseForNameDuplicates();